refactor: Integer -> Long 변경

+++ airdnb/src/main/java/team10/airdnb/accommodation/entity/Accommodation.java
@@ -43,7 +43,7 @@ public class Accommodation {
     private String name;
 
     @Column(name = "max_capacity", nullable = false)
-    private Integer maxCapacity;
+    private Long maxCapacity;
 
     @ManyToOne
     @JoinColumn(name = "accommodation_type", nullable = true)
